# projet-08-Ot-gether-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ot'gether API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get list of categories",
                "operationId": "48cfaa7170afec6cf86a79a6c0dff452",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create one new category",
                "operationId": "1045104c0f234aaa53facdf2c66bb7ba",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category create success"
                    },
                    "422": {
                        "description": "Add category error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/{category}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create one category",
                "operationId": "09943a71c052a7d746f462b79a3c199f",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "category not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete one category by id",
                "operationId": "a6b045ab91ce19f59363f5fde0aaa06b",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update one category by id",
                "operationId": "8f4dc589b179d3f5386d053aea3db97f",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "query",
                        "description": "ID of category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Update category error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/admin/events": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get list of admin events",
                "operationId": "33cffc35c432f8f628a200aec54badbf",
                "parameters": [
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Number of items per page (default is 15)",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of admin events"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Register a new event admin",
                "operationId": "88b4fae6877c237312d092bf25b4122e",
                "parameters": [
                    {
                        "name": "isIRl",
                        "in": "query",
                        "description": "Is event in Real Life",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Event title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categorie",
                        "in": "query",
                        "description": "Event category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "participant_min",
                        "in": "query",
                        "description": "Event minimal participant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "participant_max",
                        "in": "query",
                        "description": "Event maximum participant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Event start date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Event's inscription end date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Event city",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Event region",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Event address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Event registered successfully"
                    },
                    "422": {
                        "description": "Add event error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/admin/events/{event}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get one admin event",
                "operationId": "056bec188d17cecf4a0dd69ade78f98d",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the admin event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin event details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Admin event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete one admin event",
                "operationId": "97a0e03b97fdca2b3c830154b9dc5acd",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the admin event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Admin event deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Admin event not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update one admin event",
                "operationId": "0ddcbde4ad98f76d9ddaff15fc22b1b5",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the admin event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin event updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Admin event not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}